"Downgraded" post processing to 2.1.0 to fix split screen
Unity is a rollercoaster :)
Previously used "render to texture" to split cameras to different textures on canvas. This caused flickering particle systems and ocassional glitchy polygons on the viewports.
Split screen (the way we used to do it) only didn't work because apparently we had the wrong version of post processing (an earlier one works) so it is now downgraded and we use the same split screen as before.

Still needs indicators that rooms are "destroyed".